+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Potentials:
   - Added `semispring_potential`: Spring potential that is only effective for
     distances longer than the equilibrium distance.
+  - Added `wca_potential`: Lennard-Jones potential with a cut-off at the
+    potential minimum.
   - Added `wrapped_potential`: Identity wrapper for potential functor. This
     introduces custom potential in the `md` namespace, which enables arithmetic
     operations on the custom potential.